+++ release.ps1
@@ -1,6 +1,6 @@
 param(
     [ValidatePattern('^\d+\.\d+\.\d+$')]
-    [string]$Version = "1.3.4",
+    [string]$Version = "1.3.5",
     [string]$DotNet = "dotnet"
 )